pyproject.toml
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "project.settings"
addopts = "-ra --capture=no"
python_files = "tests.py test_*.py *_tests.py"

[tool.poetry]
name = "backend"
version = "0.1.0"
description = "API for Ember.js Super Rentals Tutorial"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4"
djangorestframework = "^3"
djangorestframework-jsonapi = "^5"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
openapi = "^1.1.0"
django-cors-headers = "^3.8.0"
PyYAML = "^5.4.1"
uritemplate = "^4.0.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pytest-django = "^4.4.0"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
